powerpc: allow to build big-endian binaries on little-endian host
boot_rom will be always big endian (like SLOF is).
Make the endianness of the test files configurable with
arch_CFLAGS and arch_LDFLAGS.

For the moment, this only works with big endian.

Of course, once build on a little endian host, these binaries
can be used on the little endian host to test the
big endian mode of KVM.

Signed-off-by: Laurent Vivier <lvivier@redhat.com>